Get rid of more no-longer-needed moves.
