+++ src/allmydata/web/common.py
@@ -11,7 +11,6 @@ from twisted.web import (
     server,
     template,
 )
-from twisted.web.template import tags as T
 from twisted.web.iweb import IRequest as ITwistedRequest
 from twisted.python import log
 if PY2:
@@ -508,36 +507,3 @@ def render_exception(f):
             description, status = humanize_exception(e)
             return resource.ErrorPage(status, "Error", description).render(request)
     return g
-
-
-class ReloadMixin(object):
-    REFRESH_TIME = 1*60
-
-    @template.renderer
-    def refresh(self, req, tag):
-        if self.monitor.is_finished():
-            return ""
-        # dreid suggests ctx.tag(**dict([("http-equiv", "refresh")]))
-        # but I can't tell if he's joking or not
-        tag.attributes["http-equiv"] = "refresh"
-        tag.attributes["content"] = str(self.REFRESH_TIME)
-        return tag
-
-    @template.renderer
-    def reload(self, req, tag):
-        if self.monitor.is_finished():
-            return b""
-        # url.gethere would break a proxy, so the correct thing to do is
-        # req.path[-1] + queryargs
-        ophandle = req.prepath[-1]
-        reload_target = ophandle + b"?output=html"
-        cancel_target = ophandle + b"?t=cancel"
-        cancel_button = T.form(T.input(type="submit", value="Cancel"),
-                               action=cancel_target,
-                               method="POST",
-                               enctype="multipart/form-data",)
-
-        return (T.h2("Operation still running: ",
-                     T.a("Reload", href=reload_target),
-                     ),
-                cancel_button,)
