+++ src/allmydata/web/operations.py
@@ -1,6 +1,14 @@
-
+from future.utils import PY2
 import time
-from nevow import url
+if PY2:
+    from nevow import url
+else:
+    # This module still needs porting to Python 3
+    url = None
+from twisted.web.template import (
+    renderer,
+    tags as T,
+)
 from twisted.python.failure import Failure
 from twisted.internet import reactor, defer
 from twisted.web import resource
@@ -14,11 +22,6 @@ from allmydata.web.common import (
     get_arg,
     boolean_of_arg,
     exception_to_child,
-)
-
-# Originally part of this module, so still part of its API:
-from .common import (  # noqa: F401
-    ReloadMixin,
 )
 
 MINUTE = 60
@@ -143,3 +146,36 @@ class OphandleTable(resource.Resource, service.Service):
             self.timers[ophandle].cancel()
         self.timers.pop(ophandle, None)
         self.handles.pop(ophandle, None)
+
+
+class ReloadMixin(object):
+    REFRESH_TIME = 1*MINUTE
+
+    @renderer
+    def refresh(self, req, tag):
+        if self.monitor.is_finished():
+            return ""
+        # dreid suggests ctx.tag(**dict([("http-equiv", "refresh")]))
+        # but I can't tell if he's joking or not
+        tag.attributes["http-equiv"] = "refresh"
+        tag.attributes["content"] = str(self.REFRESH_TIME)
+        return tag
+
+    @renderer
+    def reload(self, req, tag):
+        if self.monitor.is_finished():
+            return ""
+        # url.gethere would break a proxy, so the correct thing to do is
+        # req.path[-1] + queryargs
+        ophandle = req.prepath[-1]
+        reload_target = ophandle + "?output=html"
+        cancel_target = ophandle + "?t=cancel"
+        cancel_button = T.form(T.input(type="submit", value="Cancel"),
+                               action=cancel_target,
+                               method="POST",
+                               enctype="multipart/form-data",)
+
+        return (T.h2("Operation still running: ",
+                     T.a("Reload", href=reload_target),
+                     ),
+                cancel_button,)
